Day-11: Monkey in the Middle (#22)
To solve this challenge in bash, I needed to first learn about modular arithmetic to keep the item values from overflowing. Each monkey determines whether to throw an item to another monkey based on if the item value is divisible by a unique prime number. These prime numbers are multiplied together, and after an item's value was calculated like the previous challenge, the modulus of the product of all the prime numbers was used to filter any number down below the integer overflow threshold. Also, the solution to this challenge takes ~60minutes to complete, and I have included a pre-calculated log, and a flag to bypass this (and any future challenge that takes too long), within the github-action automation.